+++ packages/browser/src/index.ts
@@ -6,15 +6,22 @@ import { nextTick } from './lib/nextTick';
 import jsErrorPlugin from './plugins/jsErrorPlugin';
 import promiseErrorPlugin from './plugins/promiseErrorPlugin';
 import lifecyclePlugin from './plugins/lifecyclePlugin';
-
+import { getExtendsInfo, getStoreUserId } from './lib/utils';
 class BrowserClient extends Core<BrowserOptionType> {
     private readonly queue: Queue<BaseOptionsType>;
     protected sessionID: string;
     protected userID: string;
+    public browserOptions: BrowserOptionType;
     constructor(options) {
         super(options);
+        this.browserOptions = options;
+        this.initBrowserOptions(options);
         // 初始化队列, 用于储存上报数据结构
         this.queue = new Queue(options);
+    }
+    initBrowserOptions(options) {
+        const { enabledBasePlugins = true, enabledDeviceInfo = true } = options;
+        this.browserOptions = { ...options, enabledBasePlugins, enabledDeviceInfo }
     }
     // 初始化应用， 后期看是否需要， 因为许多参数都是传进来的， 没必要初始化， 先留个口
     async initApp() {
@@ -61,17 +68,27 @@ class BrowserClient extends Core<BrowserOptionType> {
         }
         const { title } = document;
         const { href } = window.location;
-        let deviceInfo = DeviceInfo.getDeviceInfo();
-        let deviceInfoStr = JSON.stringify(deviceInfo);
+        const { userIdentify = {}, enabledDeviceInfo } = this.browserOptions;
+        const { name: userPath, postion: userPosi } = userIdentify;
+        const user_id = getStoreUserId(userIdentify) || '';
+        if (userPath && userPosi && user_id) {
+            this.userID = user_id;
+        }
         // 在这里一会进行 公共部分的数据处理
-        return {
+        let result:BrowserReportPayloadDataType = {
             session_id: this.sessionID,
             user_id: this.userID,
             page_title: title,
             path: href,
-            deviceInfo: deviceInfoStr,
+            extendInfo: getExtendsInfo(this.context?.extendInfo || {}),
             ...data
         };
+        if (enabledDeviceInfo) {
+            let deviceInfo = DeviceInfo.getDeviceInfo();
+            let deviceInfoStr = JSON.stringify(deviceInfo);
+            result.deviceInfo = deviceInfoStr
+        }
+        return result;
     }
 }
 
@@ -79,7 +96,13 @@ const init = (options: BrowserOptionType) => {
     try {
         const client = new BrowserClient(options);
         const { plugins = [] } = options;
-        client.use([jsErrorPlugin, promiseErrorPlugin.call(client), lifecyclePlugin.call(client, options), ...plugins]);
+        const { enabledBasePlugins = true } = client.browserOptions;
+        let pluginArr = []
+        if (enabledBasePlugins) {
+            pluginArr = [jsErrorPlugin, promiseErrorPlugin.call(client)]
+        }
+        client.use([lifecyclePlugin.call(client, options), ...pluginArr, ...plugins]);
+        // client.use([jsErrorPlugin, promiseErrorPlugin.call(client), lifecyclePlugin.call(client, options), ...plugins]);
     } catch (error) {
         console.debug('===@monitor-sdk error===', error);
     }