+++ packages/browser/src/lib/utils.ts
@@ -1,0 +1,45 @@
+import { CustomerOptionType, StoreType } from '@monitor-sdk/types';
+import { getCookie, getDeepPropByDot, getStore, isFunction, isObject } from '@monitor-sdk/utils';
+
+export function getStoreUserId(userIdentify: CustomerOptionType = {}) {
+    const { name = '', postion = '', needParse = true } = userIdentify;
+    switch (postion) {
+        case StoreType.LOCAL:
+        case StoreType.SESSION:
+            return getStore(postion, name, needParse);
+        case StoreType.COOKIE:
+            return getCookie(name);
+        case StoreType.GLOBAL:
+            return getDeepPropByDot(name, window);
+        default:
+            break;
+    }
+}
+/**
+ * 获取扩展信息
+ */
+export function getExtendsInfo(data: any) {
+    try {
+        let ret = {};
+        let extendsInfo = data || {};
+        let dynamicParams;
+        if (isFunction(extendsInfo.getDynamic)) {
+            dynamicParams = extendsInfo.getDynamic(); //获取动态参数
+        }
+        //判断动态方法返回的参数是否是对象
+        if (dynamicParams && isObject(dynamicParams)) {
+            extendsInfo = { ...extendsInfo, ...dynamicParams };
+        }
+        //遍历扩展信息，排除动态方法
+        for (var key in extendsInfo) {
+            if (!isFunction(extendsInfo[key])) {
+                //排除获取动态方法
+                ret[key] = extendsInfo[key];
+            }
+        }
+        return ret;
+    } catch (error) {
+        console.log('call getExtendsInfo error', error);
+    }
+    return null;
+}