+++ packages/record/rollup.config.mjs
@@ -1,0 +1,3 @@
+import { umdPackage, iifePackage} from '../../rollup.base.config.mjs';
+
+export default [umdPackage, iifePackage];