+++ packages/console/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @monitor-sdk/console
 
+## 0.0.6
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @monitor-sdk/types@1.1.6
+    -   @monitor-sdk/utils@1.2.4
+
 ## 0.0.5
 
 ### Patch Changes