+++ packages/core/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @monitor-sdk/core
 
+## 1.1.9
+
+### Patch Changes
+
+-   新增行为录制插件
+-   Updated dependencies
+    -   @monitor-sdk/types@1.1.6
+    -   @monitor-sdk/utils@1.2.4
+
 ## 1.1.8
 
 ### Patch Changes