+++ packages/performance/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @monitor-sdk/performance
 
+## 0.1.9
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @monitor-sdk/types@1.1.6
+    -   @monitor-sdk/utils@1.2.4
+
 ## 0.1.8
 
 ### Patch Changes